Do not calculate LAST until it is needed
No significant speed-up here - it just rationalises the way LAST is
calculated. It used to be re-calculated whenever a relevant attribute
(Epoch, ObsLon, etc) was changed. But now changing an attribute just
causes it to be nullified - it is not re-calculated until it is actually
needed. Should save a few re-calculations.