New KeyMap attribute 'MapLocked' prevents new entries being added to a KeyMap.
A typical use of this to create a KeyMap, add a set of entries to it that give
default values for config parameters (etc), then set MapLocked true. Then read
values from a user-supplied config file, storing them in the KeyMap. An error
will then be reported if the config file specifies a parameter that has not been
pre-stored in the KeyMap.