Speed up access to KeyMaps that contain large numbers of entries
For instance, a speed-up of a factor of 20 for access to a KeyMap holding
30000 entries. The factor becomes larger for larger KeyMaps. The problem
was that locking a KeyMap pointer for access by the current thread
required all entries in the KeyMap to be checked to find any entries
holding AST Objects, so that these Objects could be locked too. For
KeyMaps holding few AST Object entries, this is very inefficient. The
solution was to maintain a linked list of all AST Object entries in the
KeyMap, so that they can be found quickly, without the need to check
every single entry in the keyMap.