Allow channels to read from and write to named text files
Finally got fed up with the constant need to write exactly the same
piddly little source and sink functions for use with AST Channels. So
I've added new attributes to the Channel class, called SinkFile and
SourceFile. SinkFile can be set to the path of the file to which
astWrite should write. SourceFile can be set to the path of the file
from which astRead should read.

If set, these attriobutes take precedence over the source and sink
functions specified when the channel was created.