Provide local (non-IAU) algorithm for +/- PI normalisation
Bill Joye complained that a bug I fixed last April has returned as a
result of switching to the PAL library. I had provided a different
implementation of the +/- PI angle normalisation function (palDrange)
that seems to do better for angles very close to +/- PI. Since we can't
change sofa, and PAL is supposed to be just a layer on top of sofa, I've
added a local AST function to implement the alternative algorithm. This
seems the line of least danger.