Fix bug in mathmap Gaussian noise generator
The version left by RFWS was undocumented, probably because it contains a
nasty bug that can cause seg faults.  I've done a hack that fixes this,
since I need a Gaussian noise generator for the new astRebinseq tester.
However, I've left it undocumented as I'm not at all sure it is the
correct bug fix. But it is good enough for now.