Further changes to astRebinSeq
This should fi the speed problem seen by makemap, whilst also doing
correct normalisation of the output values. The default behaviour of
astRebinSeq is no longer to conserve flux in each input array. By
default, each output value is now a weighted mean of the neighbouring
input values, without any scaling to take account of any change in pixel
size. To get flux conservation, the AST_CONSERVEFLUX flag should be
supplied when calling astRebinSeq.

OK. This is a backward incompatibility, but as sure as eggs is eggs we
are the only people to use astRebinSeq.