Fix FrameSet bug that could cause inter-Frame mappings to be inverted
If astAddFrame was called with the "iframe" argument set to
AST__ALLFRAMES, then internal Mappings within the FrameSet could be
inverted unintentionally, resulting in completely wrong WCS.

The AST__ALLFRAMES flag is only used within atlAddWcsAxis as far as I can
see. But atlAddWcsAxis is used in a number of different applications,
particularly in smurf.