Retain longitude values in the region of the pole
Previously, when transforming an (x,y) position to (long,lat), WcsMap
forced longitude to zero for positions very close to the pole. But this
causes significant round trip errors when doing pixel->sky->pixel
transformation for points very close to the pole, if the pixel size is
very small. The longitude at the pole is indeterminate, but whatever
random numerical value is returned by atan2 is no less useful (and no
more useful) than a fixed value of zero.

This was causing DS9 "compass" display to show a grossly wrong direction
for north when used with very small pixels, since it uses points at the
reference point (i.e. the pole of the FITS-WCS Native Spherical
Coordinate system) to determine the direction of north.