Make astLinearApprox more robust
Previously, a pair of LutMap describing a list of points on a raster
pattern could be deemed (incorrectly) to be linear by astLinearApprox.
This was because the test points used by astLinearApprox were on a
regular grid which just happened to hit the LutMaps in just the
correct positions to make the LutMaps appear linear.

The fix is to offset the tests points slightly from a regular grid.