Avoid creating outline polygons that have vertices on pixel edges
The polygon is now shrunk slightly, by moving each vertex by 0.01 pixels
towards the inside of the polygon. This is because rounding errors in
subsequent code may otherwise push the point into the next pixel, which
may not have valid WCS (e.g. pixels adjacent to a polar cusp in an HPX
projection).