Use global MAX, MIN and EQUAL macros instead of local versions
The basic astEQUAL macro has a fixed tolerance of 1E5*DBL__EPSILON. The
new astEQUALS macro allows the tolerance to specified. All local verions
of EQUALS have been replaced by astEQUAL or ASTEQUALS, in such a way as
to retain the original tolerances. SOme of these tolerances should be
re-visited when time allows to check that they are still appropriate.