The protected PermSplit attribute should be mutable.
It needs to be mutable so that Mappings that include PermMaps can be
simplified effectively - e.g. see the ChangePermSplit function in
fitschan.c. Making it mutable is acceptable since the PermSplit attribute
does not directly affect the nature of either the forward or the inverse
transformation of the PermMap.