Fix loading of Regions if no. of axes in base and current Frames differ
Previously used to get an error when loading Regions that had been mapped
into a space of different dimensionality to that in which the Region was
initially defined. For instamce, if a Region is created in a 4-D space,
then mapped into 2-D and dumped to disk, an error would be reported when
it was subsequently loaded back from disk.