astMapRegion now removes any unused base Frame axes
If the astMapRegion method is used to map a Region into a new Frame that
has fewer axes than the original Region, and if the inverse
transformation of the supplied Mapping does not specify a value for the
missing axes, then those axes are removed entirely from the Region.
Previously they were retained, but supplied automatically with bad
values. This affects the number of mesh points per axes for such Regions,
and so affects the accuracy of overlap determination.