Add new frame attribute Dtai
This is the offset of TAI from UTC, analogous to the attribute
Dut1 giving the offset of UT1 from UTC.  When this value is unset
it has a default value of AST__BAD and TimeMap's Transform function
will use the offset given by astDat as before.  If the new attribute
is set then it should be used for TAI-UTC and UTC-TAI conversions.

This new attribute may be useful when astDat does not return the
desired value.  Possible examples include processing an observation
which was taken since the leap second information in astDat was last
updated, or when the telescope control software used an inappropriate
TAI-UTC offset.