Check if dtai=AST__BAD before storing in CalcLAST
If dtai is the bad value, clear the Dtai attribute of tdbframe
and lastframe rather than setting it.  This is to avoid assuming that
the value stored in the dtai component will always be AST__BAD when
the attribute is unset, only that astGetDtai will return AST__BAD
in this case.