Handle simplifications that fail due to inappropriate mappings
E.g. if the simplification process assumes that an inverse transformation
is available, but doesn't actually check that it is available, then an
error may occur during simplification. Reporting an error is such cases
is bad - a clone of the supplied mapping should be returned instead. One
way to fix this is for each and every mapping class to check that the
required transformation exists before using it. But making these changes
would be laborious and error prone. So instead, just put the check in the
astSimplify_ wrapper in the base Mapping class - if a "transformation
undefined" error occurs during simplification, annull the error and
return a clone of the supplied Mapping.