Modify method polymap:polytran to cater for changes in scale
The initial guess at the best fitting polynomial used to be a unit
transformation. But if the mapping being fitted introduces a big change
in scale (e.g. if it goes from - say - mm to radians), then a unit
transformation can be a really bad guess at the best fitting
transformation. So now use an initial guess which takes account of any
overall change in scale between input and output.