New attribute FitsTol gives max allowed deviation from linearity
When writing a FrameSet out using a foreign encoding, the write operation
fails if the pixel->IWC mapping deviates from linearity by more than a
given number of pixels on any axis. This limit was previously hard-wired
at 0.1, but can now be controlled using the new FitsChan attribute
"FitsTol". If the test is passed, the linear fit to the Mapping is used
to determine the values for the FITS keywords CRPIX, CDELT, etc.