Fix bug that caused FITS string values to be padded with unneeded spaces
The recent change to use AST__DBL_DIG caused some a double value to
become much longer when converted to a string (previously it was being
rounded to a short string). The RoundFString successfully rounded the
trailing group of 9's , but then right-justiofied the string, thus
keeping its very long length. This over-flowed a buffer in smurf.

One answer is to make the buffer longer, but these extra leading spaces
are not needed, so this change modifies RoundFSTring in order to
re-instate the original behaviour (in which a short string was returned).