fitschan now puts any IWC frame between pixel and sky frames
If an IWC Frame is included in the FrameSet returned by astRead,
ensure it comes between the pixel and sky frames in the mapping chain.
Previously the order was PIXEL->SKY->IWC, Now it is PIXEL->IWC->SKY.
The inter-Frame Mappings required by the new arrangment are simpler
than for the old arrangement.