Correct docs for status argument of the TableSource function
The original intention was that the status argument for the user-supplied
TableSource function in the FitsChan class should behave like a normal
Starlink inherited status value both in C and in F77. However, a bug in
the C wrapper function means that the sense of the error value is
inverted when called from C (i.e. zero is interpreted as failure and any
non-zero value as success). But it's too late to change this now since
there may be C code out there that uses the current definition (gaia and
DS9 do for instance). So instead just document the current bahaviour.