Use numerical linearity test when checking for SIP conformity
Previously the Mapping were checked for linearity by checking that each
atomic Mapping within them was linear. But this caused some CmpMaps that
are effectively linear, but which contain grossly non-linear mappings, to
be rejected. For instance, a spherical rotation between ICRS and FK5,
projected using a TAN projection is very nearly linear on the projection
plane.