astRebinSeq: Allow the initial weight to be given for each pasted data array
astRebinSeq is used repeatedly to paste multiple input data arrays into
an output array, before finally normalising the output array. This change
allows different weights to be used for each input array, in addition to
the per-pixel weighting derived from the input Variance values. The
weight for each input pixel is the supplied initial weight divided by the
input variance

The only way to accomodate this change without changing the astRebinSeq
API was to supply the initial weight in the "params" array - not ideal
but better than changing the API.