polymap: Improve iterative inverse
This change improves the iterative inverse in cases where the linear
truncation of the PolyMap (i.e. the Mapping you get if you set all the
non-constant and non-linear coefficients in the PolyMap to zero) is far
from a unit transformation (i.e. the PolyMap introduces significant
scaling or shift).