polymap: Use Horner's method to evaluate polynomials
This was an attempt to improve the accuracy and speed of PolyMaps by
using Horner's method to evaluate each 1,2 or 3D polynomial. Particularly
the fitted inverse. However, it really didn't seem to improve anything,
which is maybe odd. The implementation is incomplete in that it does not
deal with issues arising from the different methods used by the derived
ChebyMap class (it reverts to the old method of evaluating polynomialsd
for derived classes, as it does for polynomials in more than 3
variables).

I've saved the files in this branch just in case I ever want to re-visit
the issue.