Correct buffer sizes used with AST__DBL_DIG etc
This fixes buffer overflows when writing floating point values to a char
buffer with AST__DBL_DIG significant digits. The maximum size of buffer
required is AST__DBL_DIG+6 because in addition to the AST__DBL_DIG
significant digits, there may be a decimal point and there may be an
exponent that can take up to 5 digits.