New attribute TextGapType for the plot class.
The new attribute indicates the reference position from which the gaps
specified by TextLabGap and TitleGap should be measured. By default, the
gaps are measured from the nearest edge of the bounding box containing
all other parts of hte annotated axes (excluding other textual labels).
This is the past behaviour, so no change by default. If TextGapType is
set to "plot", then the gaps are measured from the edge of the plotting
window.

This was requested by Angus Comrie of IDIA (South Africa). Who was
finding that the lower axis label (e.g. "Right Ascension") was moving up
and down depending on the size of the Y axis numerical labels (he uses
LabelUp(2)=0 so that the Y axis numerical labels are drawn with a
vertical base line and so may extend below the bottom edge of the
plotting window).