Fix spurious simplification failures
This bug could cause simplified Mappings to fail occasionally without
error (i.e. the simplified Mapping does not have the same mathematical
effect as the supplied Mapping).

The problem was that list of atomic mapping pointers passed to the
astMapMerge method of each class could potentially contain multiple
pointers to the same Mapping. The astMapMerge method for some classes of
Mapping (e.g. MatrixMap) make temporary changes to the mappings supplied
in the list. So making a change using one Mapping pointer could
potentially cause other Mappings to change, if the pointers point to the
same Mapping. This could produce unpredictable results.

This change fixes this by ensuring that deep copies are taken of any
Mappings that occur multiple times within the list of pointers sent to
the astMapMerge method.