region: Fix problem simplifying ellipses
PWD found a bug in which an ellipse was simplified to a circle that was
very different to the original ellipse. Turned out it was because the
uncertainty region in the ellipse was being misinterpreted because the
base frame in his ellipse, a simple Frame representing grid coords, did
not have a domain name set. This caused astConvert to mess up when
mapping the default uncertainty region in the ellipse, causing it to come
out much larger than it should be. This in turn caused the circle to
appear as a good fit to the ellipse.

This fix changes the way the alignment is done to avoid this.

Change Conv() so that it aligns in the current Frame of "to". Without this,
*        it is possible for uncertainty regions to be mapped incorrectly into the
*        frame of the new region when creating new regions.