mapping: Fix rebinning problem if Mapping changes handed-ness
When rebinning a data array, use the absolute value of the
determinant of the Mapping's linear approximation as the
flux conservation factor. Previously the signed determinant
was used directly. Thus negative determinants (as generated if
the input and output pixel coords have different handed-ness)
caused flux values to change sign.