Correct default CRPIX value in FitsChan
If a FITS-WCS header defines more WCS axes than pixel axes (i.e. the
header does NOT use the usual trick of defining a degenerate pixel axis
that spans only one pixel) then, internally, the FitsChan class invents
extra degenerate pixel axes itself and feeds them a value of 1.0 when
doing the forward (pixel->sky) transformation. However, the default value
for CRPIX specified in the FITS-WCS papers is 0.0, not 1.0. So it is more
appropriate for FitsChan to feed missing pixel axes a value of 0.0 rather
than 1.0. This commit fixes this (see issue "Incorrect transform from
FITS with N image axes and N+1 World axes #11" on GitHub).