Allow selected mappings to be protected from simplication
Added the RESTRICTED_SIMPLIFY and ALLOW_SIMPLIFY flags to the base
Mapping structure. If the astSimplify method is called on a Mapping for
which the RESTRICTED_SIMPLIFY flag is set, the simplification is
restricted to component Mappings that have the ALLOW_SIMPLIFY flag set.
Both of these flags are initially cleared when a Mapping is constructed.
The RESTRICTED_SIMPLIFY flag is always cleared in the Mapping returned by
astSimplify. The purpose of all this is to allow selected sections of a
complex CmpMap to be protected from simplification. These flags are
protected (i.e. not available in the public API).