mapping: Fix bug in astRebinSeq final normalisation
This fixes a bug that appeared in cases where the overlap between two
data sets is very small. In such cases it was possible for NaN values to
appear in the output array. The fix corrects the way in which the mean
weight per input pixel is calculated. It may cause some general minor
changes to the decision about which output pixels pass the "wlim"
criterion.