moc: Allow MinOrder > MaxOrder in astAddPixelMask
Previously astAddPixelMask constrained MinOrder to be less than
MaxOrder, meaning that we could not search for small features to
be included in a lower-resolution MOC.

This commit changes the behavior of astAddPixelMask to temporarily
use a maximum order of MinOrder in this case, keeping track
of the number of orders by which the MOC is being over-sampled,
similarly to the "shift" variable of the astAddRegion method
for MOCs.  When the list of cells is added to the MOC with the
IncorporateCells method, it includes the over-sampling number
in its "shift" variable.  The logic for negative shifts
from astAddRegion was applied here also.