moc: Attempt to avoid RegBaseMesh getting trapped
The boundary tracing method could previously enter a loop
for some MOC structures (typically at the edges of SCUBA-2 maps)
triggering the "Re-drawing previous..." error message.  Therefore
record on which path a corner has been seen (as "path_start")
to avoid tracing it multiple times.  However this could lead
to becoming stuck (and the "Next perimeter corner is undefined..."
error message).  In this case, mark the current corner as not
part of the path (with "path_backtrack") and go back to the
start.  Finally, in order to fill the "meshdist" array
correctly it seemed to be necessary to record separately
a second "dist" value if a corner appeared in two paths.