moc: Reverse backtracking direction in RegBaseMesh
Now when a dead end is reached when searching for the next corner
along a boundary path, step back to the previous corner rather than
beginning the path again.  This is to accommodate situations where
a corner is initially found in an invalid path, but is actually
part of the correct path -- previously it would not be found having
been marked as not part of the current path.  This approach
does involve tracking more information (in the Corner object)
but is faster due to not having to repeatedly trace the path
from the start.  Added a note to the news file as the routine
now seems to be able to trace more MOC structures than before.