moc: Process isolated cells last in RegBaseMesh
This adds an option to the MakeCorners method to have it attempt
to mark isolated corners and cells.  Then RegBaseMesh can process
these separately when tracing the region boundary by not beginning
a path at one of these corners.  Then for any such cell which was
not otherwise drawn, a path around its corners is added to the
end of the mesh distance array.

This is intended to attempt to avoid problems caused by isolated
cells in some MOC structures, such as where there is a checkerboard
pattern of single cells and holes.