mapping: Correct bug in astRate algorithm
For some edge case Mappings, the algorithm used to find the best interval
for determing the gradient fails and uses a ludicrously large or small
interval size, resulting in an incorrect estimate of the gradient. This
commit fixes this bug by terminating the search for the best interval
size when the first local minimum is found (previously it looked for the
global minimum).