Makefile: attempt to detect errors running astbad
Run astbad in separate assignment commands to allow the exit status
of each to be bad on failure to run it.  This is to allow make to
detect this problem, otherwise an AST_PAR file can be generated
with blank values for the corresponding parameters.  This can happen,
for example, on macOS, if trying to link to a copy of libyaml
of which the identification name has been changed.