Makefile: pass FFLAGS to ast_tester
The flag "-fallow-argument-mismatch" is required for ast_tester on
some systems but not supported on others.  The build system already
determines whether the flag is supported and includes it, along with
-I options, in STAR_FFLAGS if so.  Therefore provide a modified
copy of this variable, STAR_FFLAGS_TEST with the -I options removed.
Then the "make test" rule can pass this as FFLAGS to ast_tester.
Finally update ast_tester to include "-fno-range-check" even
when some FFLAGS are already given.